Move CollectRunfolderSamples class to toolbox, move dx run command functions to build_dx_commands script, create classes per run type to make commands generation easier to follow, give dev run uploads the same behaviour as tso runs (uploads BCLs), move PipelineEmails to separate script
